requirements.txt
django
djangorestframework
markdown
django-filter
requests
djangorestframework_simplejwt